full unit test for time travel component
