+++ src/app/components/stateChange.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import {deleteHtmlElement, fixState,componentChangedState} from './helperFunctions'
-// import {componentChangedState} from './helperFunctions'
+import ComponentsList from './componentsList'
 
 export default class stateChange extends Component {
     constructor(props){
@@ -29,7 +29,7 @@ export default class stateChange extends Component {
             <h2 className='graph-title'>State Diff</h2>
               <div className="state-changes">
                   <div className="wrapper">
-                      <ComponentsList2 
+                      <ComponentsList 
                           components={[current]} 
                           onChange={(selectedcomponents) => this.setState({selectedcomponents})}
                           selectedcomponents={selectedcomponents} 
@@ -46,117 +46,4 @@ export default class stateChange extends Component {
         }
     }
 }
-
-const ComponentsList2 = ({ components, selectedcomponents, onChange, isFirst, debug }) => {
- 
-    const handleCheckboxClicked = (selectedcomponentId) => {
-  
-      if(selectedcomponents[selectedcomponentId]){
-        delete selectedcomponents[selectedcomponentId];
-      } else {
-        selectedcomponents[selectedcomponentId] = {}
-        
-      }
-      
-      onChange(selectedcomponents)
-    }
-    
-    const handleSubcomponentsListChange = (componentId, subSelections) => {
-      selectedcomponents[componentId] = subSelections;
-      onChange(selectedcomponents);
-    }
-    
-    let counter=0
-    return (
-      <div>
-        {components.map(component => (
-            <ul className={isFirst && "firstUL"}>
-              <Checkbox2 
-                state={component.state}
-                props={component.props}
-                selected={selectedcomponents[component.name]} 
-                label={component.name} 
-                onChange={() => {handleCheckboxClicked(component.name)}}
-              />
-              {(component.children.length > 0 && selectedcomponents[component.name]) &&
-                <ComponentsList2 
-                  key={counter+=100}
-                  components={component.children} 
-                  selectedcomponents={selectedcomponents[component.name]} 
-                  onChange={(subSelections) => handleSubcomponentsListChange(component.name, subSelections)}
-                />
-              }
-            </ul>
-          )
-        )}
-      </div>
-    )
-  }
-
-class Checkbox2 extends Component {
-    constructor(props){
-        super(props)
-        this.state={
-            didmount:true
-        }
-    }
-    componentDidMount(){
-        // if(this.state.didmount){
-            this.props.onChange(!this.props.selected)
-        //     this.setState({didmount:false})
-        // }
-      }
-      creatlistState(value){
-        const temp=[]
-        if(value !== null){
-          if(!Array.isArray(value) && value!== null && value !== undefined && Object.getPrototypeOf(value).isPrototypeOf(Object)){
-          for(let key in value){
-            if(typeof value[key] ==='object' && value[key] !=null) temp.push(this.creatlistState(value[key]))
-            if(value[key]===null || value[key]=='null' || this.props.label=='UserTable') temp.push(<li key={'state' + temp.length}>{key}: null</li>)
-            if(typeof value[key] !=='object') temp.push(<li key={'state'+temp.length}>{key}:<span className='span'>{`${value[key]}`}</span> </li>)
-          }
-          return temp
-          }else if(Array.isArray(value)){
-              for(let i=0;i<value.length;i++){
-                if(typeof value[i] ==='object') temp.push(this.creatlistProps(value[i]))
-              }
-            }
-            return temp
-      }
-      return temp
-    }
-    creatlistProps(value){
-      const temp = []
-        if(value !== null){
-          if(!Array.isArray(value)){
-            for(let key in value){
-              if(typeof value[key] ==='object' && value[key] !=null) temp.push(this.creatlistProps(value[key]))
-              if(value[key]===null || value[key]=='null') temp.push(<li key={'props'+temp.length}>{key}:</li>)
-              if(typeof value[key] !=='object'  )temp.push(<li key={'props'+temp.length}>{`${key}`}: <span className="span">{`${value[key]}`}</span></li>)
-            }
-            return temp
-          }else{
-            for(let i=0;i<value.length;i++){
-              if(typeof value[i] ==='object') temp.push(this.creatlistProps(value[i]))
-              // if(value[i]===null) temp.push(<li key={'props'+temp.length}>{key}: {'null'}</li>)
-              // else temp.push(<li key={'props'+temp.length}>{key}: {`${value[i]}`}</li>)
-            }
-          }
-          return temp
-    }
-    return temp
-  }
-    render(){
-      
-        const state= (this.props.state !==null)? this.creatlistState(this.props.state):[]
-        const props= (this.props.props !==null)? this.creatlistProps(this.props.props):[]
-        return (
-          <div className="checkbox">  
-            <div className="checkbox__label"  onClick={() => this.props.onChange(!this.props.selected)}>{this.props.label}</div>
-            <li><span className='stateChangeTitles'> State </span> {(state==null ||  state ==='null')?'':state}</li>
-            <li><span className='stateChangeTitles'> Props </span> {(props==null ||  state ==='null')?'':props}</li>
-          </div>
-        )
-    }
-  }
